+++ tests/codegen/call_nsecs_boot.cpp
@@ -1,0 +1,14 @@
+#include "common.h"
+
+namespace bpftrace {
+namespace test {
+namespace codegen {
+
+TEST(codegen, call_nsecs_boot)
+{
+  test("k:f { @x = nsecs(boot); }", NAME);
+}
+
+} // namespace codegen
+} // namespace test
+} // namespace bpftrace